Solve best regression for L-infinite norm (#351)
Related to #341

Existing approach is not quite optimized in terms of finding **what is
the best line to fit the data**?

This new approach will solve for the optimal results.

However, unlike l2 norm which has closed form solutions, solving for
l-infinite norm requires optimization steps.

We only do 8 steps, and will fallback to original bit-packing if we
don't find a good enough line.